[DDW-582] Removes vertical menu
