+++ source/renderer/app/components/wallet/layouts/WalletWithNavigation.js
@@ -1,7 +1,6 @@
 // @flow
 import React, { Component } from 'react';
 import type { Node } from 'react';
-import classnames from 'classnames';
 import { observer } from 'mobx-react';
 import WalletNavigation from '../navigation/WalletNavigation';
 import styles from './WalletWithNavigation.scss';
@@ -10,22 +9,12 @@ type Props = {
   children?: Node,
   isActiveScreen: Function,
   onWalletNavItemClick: Function,
-  isSettingsPage: boolean,
 };
 
 @observer
 export default class WalletWithNavigation extends Component<Props> {
   render() {
-    const {
-      children,
-      isActiveScreen,
-      onWalletNavItemClick,
-      isSettingsPage,
-    } = this.props;
-    const pageStyles = classnames([
-      styles.page,
-      isSettingsPage ? styles.settingsTabPage : null,
-    ]);
+    const { children, isActiveScreen, onWalletNavItemClick } = this.props;
     return (
       <div className={styles.component}>
         <div className={styles.navigation}>
@@ -34,7 +23,7 @@ export default class WalletWithNavigation extends Component<Props> {
             onNavItemClick={onWalletNavItemClick}
           />
         </div>
-        <div className={pageStyles}>{children}</div>
+        <div className={styles.page}>{children}</div>
       </div>
     );
   }
