+++ source/renderer/app/components/wallet/settings/ChangeSpendingPasswordDialog.js
