+++ source/renderer/app/components/wallet/settings/DeleteWalletButton.js
