+++ source/renderer/app/components/wallet/settings/DeleteWalletButton.scss
