+++ source/renderer/app/components/wallet/settings/DeleteWalletConfirmationDialog.js
