+++ source/renderer/app/components/wallet/settings/DeleteWalletConfirmationDialog.scss
