+++ source/renderer/app/components/wallet/settings/WalletSettings.scss
