--- source/renderer/app/components/wallet/settings/WalletSettingsLayout.scss
@@ -1,24 +1,0 @@
-.component {
-  background-color: var(--theme-settings-body-background-color);
-  display: flex;
-  height: 100%;
-  padding: 0 20px;
-}
-
-.settingsPaneWrapper {
-  flex: 1;
-  overflow-x: hidden;
-  overflow-y: overlay;
-  padding: 20px 20px 20px 0;
-  &::-webkit-scrollbar-button {
-    height: 7px;
-    display: block;
-  }
-}
-
-.settingsPane {
-  background-color: var(--theme-settings-pane-background-color);
-  border: var(--theme-settings-pane-border);
-  border-radius: 4px;
-  overflow: visible;
-}
