--- source/renderer/app/components/wallet/settings/WalletSettingsMenu.scss
@@ -1,8 +1,0 @@
-.component {
-  background-color: var(--theme-settings-menu-box-background-color);
-  border: var(--theme-settings-menu-box-border);
-  border-radius: 4px;
-  margin: 20px 0;
-  padding: 16px 0;
-  width: 240px;
-}
