--- source/renderer/app/components/wallet/settings/WalletSettingsMenuItem.scss
@@ -1,32 +1,0 @@
-.component {
-  color: var(--theme-settings-menu-item-text-color);
-  display: block;
-  font-family: var(--font-regular);
-  font-size: 16px;
-  line-height: 1.38;
-  overflow: hidden;
-  padding: 14px 10px 14px 20px;
-  text-align: left;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
-}
-
-.active {
-  color: var(--theme-settings-menu-item-text-color-active);
-  background-color: var(--theme-settings-menu-item-background-color-active);
-  border-left: 4px solid
-    var(--theme-settings-menu-item-left-border-color-active);
-  font-family: var(--font-bold);
-  padding-left: 16px;
-}
-
-.enabled {
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.disabled {
-  color: var(--theme-settings-menu-item-text-color-disabled);
-}
