+++ source/renderer/app/components/wallet/utxo/WalletUtxo.scss
