+++ source/renderer/app/components/wallet/utxo/WalletUtxoCursor.scss
