+++ source/renderer/app/components/wallet/utxo/WalletUtxoTick.scss
