+++ source/renderer/app/components/wallet/utxo/WalletUtxoTooltip.scss
