+++ storybook/stories/DeleteWalletConfirmationDialog.stories.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import StoryDecorator from './support/StoryDecorator';
-import DeleteWalletConfirmationDialog from '../../source/renderer/app/components/wallet/settings-general/DeleteWalletConfirmationDialog';
+import DeleteWalletConfirmationDialog from '../../source/renderer/app/components/wallet/settings/DeleteWalletConfirmationDialog';
 
 storiesOf('DeleteWalletConfirmationDialog', module)
   .addDecorator(story => <StoryDecorator>{story()}</StoryDecorator>)
