+++ storybook/stories/WalletScreens.stories.js
@@ -28,9 +28,8 @@ import WalletSummary from '../../source/renderer/app/components/wallet/summary/W
 import WalletSendForm from '../../source/renderer/app/components/wallet/WalletSendForm';
 import WalletReceive from '../../source/renderer/app/components/wallet/receive/WalletReceive';
 import WalletTransactionsList from '../../source/renderer/app/components/wallet/transactions/WalletTransactionsList';
-
-import WalletScreensGeneralSettings from './WalletScreens-GeneralSettings.stories';
-import WalletScreensSettingsUtxo from './WalletScreens-SettingsUtxo.stories';
+import WalletScreensSettings from './WalletScreens-Settings.stories';
+import WalletScreensUtxo from './WalletScreens-Utxo.stories';
 
 /* eslint-disable consistent-return */
 storiesOf('WalletScreens', module)
@@ -163,5 +162,5 @@ storiesOf('WalletScreens', module)
     />
   ))
 
-  .add('Settings - General', WalletScreensGeneralSettings)
-  .add('Settings - Wallet UTXO distribution', WalletScreensSettingsUtxo);
+  .add('Settings - General', WalletScreensSettings)
+  .add('Settings - Wallet UTXO distribution', WalletScreensUtxo);
